管理+评论+其他

+++ app/src/main/java/com/miaofen/xiaoying/common/data/bean/response/DetailsResponse.kt
@@ -69,7 +69,7 @@ data class DetailsResponse(
         var personalSignature: String?,
         var avatarUrl: String?,
         var nickName: String?,
-        var userId: Int?
+        var userId: Long?
     )
 
     data class PlanTagsBean(
